Added message confirm address reception (change not used, under ifdef)
The main reason for this unused update was to introduce a message so that the controller can say to the Expansion
"I sent you a new address, you answer me back using this correct address... I got that, now you can leave the other expansion to have an address too".
This is good however in case an expansion with this change is plugged into a system with a controller without this it would have required an expansion update.
The same has been achieved with a simpler changes only in the controller.